Remove the old serializers (needs snapshot)
