+++ src/libsyntax/ast_util.rs
@@ -406,6 +406,8 @@ fn dtor_dec() -> fn_decl {
 // ______________________________________________________________________
 // Enumerating the IDs which appear in an AST
 
+#[auto_serialize]
+#[auto_deserialize]
 #[auto_serialize2]
 #[auto_deserialize2]
 type id_range = {min: node_id, max: node_id};
