+++ src/libsyntax/parse/token.rs
@@ -2,6 +2,8 @@ use util::interner;
 use util::interner::interner;
 use std::map::HashMap;
 
+#[auto_serialize]
+#[auto_deserialize]
 #[auto_serialize2]
 #[auto_deserialize2]
 enum binop {
@@ -17,6 +19,8 @@ enum binop {
     SHR,
 }
 
+#[auto_serialize]
+#[auto_deserialize]
 #[auto_serialize2]
 #[auto_deserialize2]
 enum token {
@@ -75,6 +79,8 @@ enum token {
     EOF,
 }
 
+#[auto_serialize]
+#[auto_deserialize]
 #[auto_serialize2]
 #[auto_deserialize2]
 /// For interpolation during macro expansion.
