+++ src/libsyntax/syntax.rc
@@ -128,9 +128,9 @@ mod ext {
     mod ident_to_str;
     #[legacy_exports]
     mod log_syntax;
-    #[cfg(stage0)]
     #[legacy_exports]
     mod auto_serialize;
+    #[cfg(stage0)]
     #[legacy_exports]
     mod auto_serialize2;
     #[legacy_exports]
